Agrego casos de prueba y docstrings
