+++ series_tiempo_ar_api/apps/metadata/tests/view_tests.py
@@ -20,3 +20,29 @@ class ViewTests(TestCase):
         self.assertIn('limit', response_json)
         self.assertIn('count', response_json)
         self.assertIn('offset', response_json)
+
+
+class DatasetSourceTests(TestCase):
+
+    def test_run(self):
+        test_source = 'test_source'
+        search_mock = mock.MagicMock()
+        search_mock.aggregations.results.buckets = [{'key': test_source}]
+
+        with mock.patch.object(Search, 'execute', return_value=search_mock):
+            response = self.client.get(reverse('api:metadata:dataset_source'))
+            response_sources = json.loads(response.content)['data']
+            self.assertIn(test_source, response_sources)
+
+
+class FieldUnitTests(TestCase):
+
+    def test_run(self):
+        test_unit = 'test_unit'
+        search_mock = mock.MagicMock()
+        search_mock.aggregations.results.buckets = [{'key': test_unit}]
+
+        with mock.patch.object(Search, 'execute', return_value=search_mock):
+            response = self.client.get(reverse('api:metadata:field_units'))
+            response_sources = json.loads(response.content)['data']
+            self.assertIn(test_unit, response_sources)
