+++ series_tiempo_ar_api/apps/metadata/views.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from django.http import JsonResponse
 
-from series_tiempo_ar_api.apps.metadata.queries.dataset_source import query_field_terms
+from series_tiempo_ar_api.apps.metadata.queries.query_terms import query_field_terms
 from series_tiempo_ar_api.apps.metadata.queries.query import FieldSearchQuery
